% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Add raku engine to knitr.}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Not used.}

\item{pkgname}{Not used.}
}
\value{
Nothing (invisible)
}
\description{
Invoke the raku engine like so:
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{```\{raku\}
...
```
}\if{html}{\out{</div>}}
}
\author{
fnaufel
}
