% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{rraku}
\alias{rraku}
\title{Add raku engine to knitr.}
\description{
Invoke the raku engine like so:
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{```\{raku\}
...
```
}\if{html}{\out{</div>}}
}
\author{
fnaufel
}
